fix: onboarding process checker added

+++ Includes/Admin/AdminHooks.php
@@ -52,6 +52,10 @@ class AdminHooks {
 	 */
 
 	public function sgsb_redirect_initial_setup() {
+		$ini_setup_completion = get_option( 'sgsb_ini_completion' );
+		if ( $ini_setup_completion ) {
+			return;}
+
 		if ( get_option( 'storegrowth_activation_redirect', false ) ) {
 			delete_option( 'storegrowth_activation_redirect' );
 			if ( wp_safe_redirect( admin_url( 'admin.php?page=sgsb-modules#/ini-setup' ) ) ) {
